Move the remainder of the code from talent.js into separate files
