+++ scripts/spellbook.js
@@ -1,32 +1,7 @@
-import { KEY as MODULE_NAME, log, CUSTOM_SHEETS } from './module.js';
-import { localise, getOrderName } from './utils.js';
+import { CUSTOM_SHEETS } from './module.js';
+import { getOrderName, localise } from './utils.js';
 
-const POWER_SPECIALTIES = [
-    'Chronopathy',
-    'Metamorphosis',
-    'Pyrokinesis',
-    'Resopathy',
-    'Telekinesis',
-    'Telepathy'
-];
-
-export const registerWithLibWrapper = function() {
-    libWrapper.register(MODULE_NAME, 'game.dnd5e.applications.actor.ActorSheet5eCharacter.prototype._prepareSpellbook', getSpellbook, 'WRAPPER');
-    libWrapper.register(MODULE_NAME, 'game.dnd5e.applications.item.AbilityUseDialog._createSpellSlotOptions', createSpellSlotOptions, 'WRAPPER');
-};
-
-export const setupPowerSpecialties = function() {
-    CONFIG.DND5E.spellPreparationModes['talent'] = localise("SpellPrepTalent");
-    CONFIG.DND5E.spellUpcastModes.push('talent');
-
-    POWER_SPECIALTIES.forEach(s => {
-        CONFIG.DND5E.spellSchools[s] = localise(`PowerSpecialty.${s}`);
-    });
-}
-
-//#region Wrappers
-
-function getSpellbook(wrapped, ...args) {
+export function getSpellbook(wrapped, ...args) {
     let nonTalentData = getNonTalentSpells(...args);
     let book = wrapped(...nonTalentData);
     let talentBook = prepareTalentSpellbook(...args);
@@ -34,25 +9,14 @@ function getSpellbook(wrapped, ...args) {
     return book;
 }
 
-function createSpellSlotOptions(wrapped, ...args) {
-    log.debug("createSpellSlotOptions", ...args);
-    let result = createPowerOrderOptions(...args).concat(wrapped(...args));
-    log.debug("getPowerData result: ", result);
-    return result;
-}
-
-//#endregion
-
-//#region Spellbook
-
 /**
  * @param {object} context  Sheet rendering context data being prepared for render.
  * @param {object[]} spells Spells to be included in the spellbook.
  * @returns {object[]}      Spellbook sections in the proper order.
  */
 function getNonTalentSpells(context, spells) {
-    let noPowers = spells.filter(s => s.system.preparation.mode != 'talent')
-    return [ context, noPowers ];
+    let noPowers = spells.filter(s => s.system.preparation.mode != 'talent');
+    return [context, noPowers];
 }
 
 /**
@@ -62,8 +26,8 @@ function getNonTalentSpells(context, spells) {
  */
 function prepareTalentSpellbook(context, spells) {
     const levels = context.actor.system.spells;
-    const powerbook = {}
-    const infinite = "&infin;"
+    const powerbook = {};
+    const infinite = "&infin;";
 
     const registerSection = (sl, lvl, label) => {
         powerbook[lvl] = {
@@ -76,7 +40,7 @@ function prepareTalentSpellbook(context, spells) {
             uses: infinite,
             slots: infinite,
             override: 0,
-            dataset: { type: "spell", level: lvl, "preparation.mode": "talent"},
+            dataset: { type: "spell", level: lvl, "preparation.mode": "talent" },
             prop: sl
         };
     };
@@ -84,9 +48,9 @@ function prepareTalentSpellbook(context, spells) {
     // Talents have power orders from 1 to 6.
     if (levels.talent) {
         const maxOrder = Array.fromRange(7).reduce((max, i) => {
-            if ( i === 0 ) return max;
+            if (i === 0) return max;
             const level = levels[`spell${i}`];
-            if ( (level.max || level.override) && (i > max)) max = i;
+            if ((level.max || level.override) && (i > max)) max = i;
             return max;
         }, 0);
 
@@ -101,7 +65,7 @@ function prepareTalentSpellbook(context, spells) {
     spells.forEach(spell => {
         const mode = spell.system.preparation.mode || "prepared";
         let lvl = spell.system.level || 0;
-        const sl = `spell${lvl}`
+        const sl = `spell${lvl}`;
 
         if (mode === 'talent') {
             if (!powerbook[lvl]) {
@@ -117,7 +81,7 @@ function prepareTalentSpellbook(context, spells) {
     return sorted;
 }
 
-export const renameSpellbookHeadings = function(sheet, html, actor) {
+export const renameSpellbookHeadings = function (sheet, html, actor) {
     const orderSuffix = localise("PowerOrderSuffix");
     const powerSpecialty = localise("PowerSpecialty.Header");
     const powerUsage = localise("PowerUsage");
@@ -132,44 +96,11 @@ export const renameSpellbookHeadings = function(sheet, html, actor) {
         headerLabels.children(".items-header-range").prop("title", powerRange);
         headerLabels.children(".items-header-usage").prop("title", powerUsage);
 
-        html.find(".tab.spellbook ul.item-list li.item div.spell-comps").hide()
+        html.find(".tab.spellbook ul.item-list li.item div.spell-comps").hide();
     } else {
         let orderSections = html.find(`.tab.spellbook div.item-name h3:contains("${orderSuffix}")`).parent();
         orderSections.siblings('.spell-school').text(powerSpecialty);
         orderSections.siblings('.spell-action').text(powerUsage);
         orderSections.siblings('.spell-target').text(powerTarget);
     }
-}
-    
-
-//#endregion
-
-/**
- * 
- * @param {dnd5e.documents.Actor5e} actor   The actor with spell slots.
- * @param {number} order    The minimum order
- * @returns {object[]}      Array of spell slot select options.
- * @private
- */
-function createPowerOrderOptions(actor, order) {
-    if (actor.classes.talent === undefined || order === 1) {
-        return [];
-    }
-
-    // Determine which orders are feasible
-    let maxOrder = Math.ceil(actor.classes.talent.system.levels / 4) + 1;
-    const powerOrders = Array.fromRange(maxOrder+1).reduce((arr, i) => {
-        if (i < order) return arr;
-        const label = getOrderName(i);
-        arr.push({
-            key: `power${i}`,
-            level: i,
-            label: label,
-            canCast: true,
-            hasSlots: true
-        });
-        return arr;
-    }, []);
-
-    return powerOrders;
-}
+};
